scrape_data/requirements.txt
igdb_api_v4==0.3.3
protobuf==5.28.2
requests
cloudscraper
tenacity